requirements.txt
praw
vaderSentiment
pandas
streamlit
psycopg2-binary
python-dotenv
matplotlib
wordcloud